requirements/base.txt
selenium==3.14.0
six==1.12.0
path.py==7.2
requests==2.22.0
nose==1.3.7
Paver==1.2.4
bok-choy==0.9.0
needle==0.5.0
pycodestyle==2.5.0
pylint==1.7.6
edx-lint==1.3.0
edx-opaque-keys==1.0.0
edx-rest-api-client==1.9.2
ddt==1.1.1
mock==1.0.1
temp-mail==0.2
python-guerrillamail==0.2.0
